+++ sylt-compiler/src/compiler.rs
@@ -4,17 +4,17 @@ use std::path::{Path, PathBuf};
 use std::rc::Rc;
 use sylt_common::error::Error;
 use sylt_common::prog::Prog;
-use sylt_common::{Block, Op, RustFunction, Type, Value};
-use sylt_parser::expression::ComparisonKind;
+use sylt_common::{Op, RustFunction, Type, Value};
 use sylt_parser::statement::NameIdentifier;
 use sylt_parser::{
     Context as ParserContext,
-    Assignable, AssignableKind, Expression, ExpressionKind, Identifier, Op as ParserOp,
-    Span, Statement, StatementKind, Type as ParserType, TypeKind, VarKind, AST,
+    Assignable, AssignableKind, Identifier, Span,
+    StatementKind, Type as ParserType, TypeKind, VarKind, AST,
 };
 
 mod typechecker;
 mod dependency;
+mod bytecode;
 
 type VarSlot = usize;
 
@@ -88,7 +88,7 @@ impl Upvalue {
 
 #[derive(Debug, Copy, Clone)]
 struct Context {
-    block_slot: BlockID,
+    // block_slot: BlockID,
     namespace: NamespaceID,
     scope: usize,
     frame: usize,
@@ -98,7 +98,7 @@ impl Context {
     fn from_namespace(namespace: NamespaceID) -> Self {
         Self {
             namespace,
-            block_slot: 0,
+            // block_slot: 0,
             scope: 0,
             frame: 0,
         }
@@ -149,12 +149,10 @@ impl Frame {
 }
 
 struct Compiler {
-    blocks: Vec<Block>,
     namespace_id_to_path: HashMap<NamespaceID, PathBuf>,
 
     namespaces: Vec<Namespace>,
 
-    loops: Vec<LoopFrame>,
     frames: Vec<Frame>,
     functions: HashMap<String, (usize, RustFunction, Type)>,
 
@@ -195,12 +193,9 @@ macro_rules! error_no_panic {
 impl Compiler {
     fn new() -> Self {
         Self {
-            blocks: Vec::new(),
-
             namespace_id_to_path: HashMap::new(),
             namespaces: Vec::new(),
 
-            loops: Vec::new(),
             frames: Vec::new(),
             functions: HashMap::new(),
 
@@ -211,20 +206,6 @@ impl Compiler {
             constants: Vec::new(),
 
             values: HashMap::new(),
-        }
-    }
-
-    fn push_frame_and_block(&mut self, ctx: Context, name: &str, span: Span) -> Context {
-        let file_as_path = PathBuf::from(self.file_from_namespace(ctx.namespace));
-
-        let block = Block::new(&name, ctx.namespace, &file_as_path);
-        self.blocks.push(block);
-
-        self.frames.push(Frame::new(name, span));
-        Context {
-            block_slot: self.blocks.len() - 1,
-            frame: self.frames.len() - 1,
-            ..ctx
         }
     }
 
@@ -266,296 +247,6 @@ impl Compiler {
         Op::Constant(slot)
     }
 
-    fn add_op(&mut self, ctx: Context, span: Span, op: Op) -> usize {
-        self.blocks
-            .get_mut(ctx.block_slot)
-            .expect("Invalid block id")
-            .add(op, span.line)
-    }
-
-    fn patch(&mut self, ctx: Context, ip: usize, op: Op) {
-        self.blocks
-            .get_mut(ctx.block_slot)
-            .expect("Invalid block id")
-            .ops[ip] = op;
-    }
-
-    fn next_ip(&mut self, ctx: Context) -> usize {
-        self.blocks
-            .get_mut(ctx.block_slot)
-            .expect("Invalid block id")
-            .curr()
-    }
-
-    fn emit_pop_until_size(&mut self, ctx: Context, span: Span, target_size: usize) {
-        let vars: Vec<_> = self.frames[ctx.frame].variables.iter().skip(target_size).rev().cloned().collect();
-        for var in &vars {
-            if var.captured {
-                self.add_op(ctx, span, Op::PopUpvalue);
-            } else {
-                self.add_op(ctx, span, Op::Pop);
-            }
-        }
-    }
-
-    fn pop_until_size(&mut self, ctx: Context, span: Span, target_size: usize) {
-        self.emit_pop_until_size(ctx, span, target_size);
-        self.frames[ctx.frame].variables.truncate(target_size);
-    }
-
-    fn assignable(&mut self, ass: &Assignable, ctx: Context) -> Option<usize> {
-        use AssignableKind::*;
-
-        match &ass.kind {
-            Read(ident) => {
-                return self.read_identifier(&ident.name, ass.span, ctx, ctx.namespace);
-            }
-            Call(f, expr) => {
-                self.assignable(f, ctx);
-                for expr in expr.iter() {
-                    self.expression(expr, ctx);
-                }
-                self.add_op(ctx, ass.span, Op::Call(expr.len()));
-            }
-            ArrowCall(pre, f, expr) => {
-                self.expression(pre, ctx);
-                self.assignable(f, ctx);
-                self.add_op(ctx, ass.span, Op::Swap);
-                for expr in expr.iter() {
-                    self.expression(expr, ctx);
-                }
-                self.add_op(ctx, ass.span, Op::Call(expr.len() + 1));
-            }
-            Access(a, field) => {
-                if let Some(namespace) = self.assignable(a, ctx) {
-                    return self.read_identifier(&field.name, field.span, ctx, namespace);
-                } else {
-                    let slot = self.string(&field.name);
-                    self.add_op(ctx, field.span, Op::GetField(slot));
-                }
-            }
-            Index(a, b) => {
-                self.assignable(a, ctx);
-                if let ExpressionKind::Int(n) = b.kind {
-                    self.add_op(ctx, ass.span, Op::GetConstIndex(n));
-                } else {
-                    self.expression(b, ctx);
-                    self.add_op(ctx, ass.span, Op::GetIndex);
-                }
-            }
-            Expression(expr) => {
-                self.expression(expr, ctx);
-            }
-        }
-        None
-    }
-
-    fn un_op(&mut self, a: &Expression, ops: &[Op], span: Span, ctx: Context) {
-        self.expression(&a, ctx);
-        for op in ops {
-            self.add_op(ctx, span, *op);
-        }
-    }
-
-    fn bin_op(&mut self, a: &Expression, b: &Expression, ops: &[Op], span: Span, ctx: Context) {
-        self.expression(&a, ctx);
-        self.expression(&b, ctx);
-        for op in ops {
-            self.add_op(ctx, span, *op);
-        }
-    }
-
-    fn push(&mut self, value: Value, span: Span, ctx: Context) {
-        let value = self.constant(value);
-        self.add_op(ctx, span, value);
-    }
-
-    fn expression(&mut self, expression: &Expression, ctx: Context) {
-        use ComparisonKind::*;
-        use ExpressionKind::*;
-
-        match &expression.kind {
-            Get(a) => {
-                self.assignable(a, ctx);
-            }
-
-            TypeConstant(ty) => {
-                let resolved_ty = self.resolve_type(ty, ctx);
-                let ty_constant = self.constant(Value::Ty(resolved_ty));
-                self.add_op(ctx, expression.span, ty_constant);
-            }
-
-            Add(a, b) => self.bin_op(a, b, &[Op::Add], expression.span, ctx),
-            Sub(a, b) => self.bin_op(a, b, &[Op::Sub], expression.span, ctx),
-            Mul(a, b) => self.bin_op(a, b, &[Op::Mul], expression.span, ctx),
-            Div(a, b) => self.bin_op(a, b, &[Op::Div], expression.span, ctx),
-
-            Comparison(a, cmp, b) => match cmp {
-                Equals => self.bin_op(a, b, &[Op::Equal], expression.span, ctx),
-                NotEquals => self.bin_op(a, b, &[Op::Equal, Op::Not], expression.span, ctx),
-                Greater => self.bin_op(a, b, &[Op::Greater], expression.span, ctx),
-                GreaterEqual => self.bin_op(a, b, &[Op::Less, Op::Not], expression.span, ctx),
-                Less => self.bin_op(a, b, &[Op::Less], expression.span, ctx),
-                LessEqual => self.bin_op(a, b, &[Op::Greater, Op::Not], expression.span, ctx),
-                Is => self.bin_op(a, b, &[Op::Is], expression.span, ctx),
-                In => self.bin_op(a, b, &[Op::Contains], expression.span, ctx),
-            }
-
-            AssertEq(a, b) => self.bin_op(a, b, &[Op::Equal, Op::Assert], expression.span, ctx),
-
-            Neg(a) => self.un_op(a, &[Op::Neg], expression.span, ctx),
-
-            And(a, b) => {
-                self.expression(a, ctx);
-
-                self.add_op(ctx, expression.span, Op::Copy(1));
-                let jump = self.add_op(ctx, expression.span, Op::Illegal);
-                self.add_op(ctx, expression.span, Op::Pop);
-
-                self.expression(b, ctx);
-
-                let op = Op::JmpFalse(self.next_ip(ctx));
-                self.patch(ctx, jump, op);
-            }
-            Or(a, b) => {
-                self.expression(a, ctx);
-
-                self.add_op(ctx, expression.span, Op::Copy(1));
-                let skip = self.add_op(ctx, expression.span, Op::Illegal);
-                let jump = self.add_op(ctx, expression.span, Op::Illegal);
-
-                let op = Op::JmpFalse(self.next_ip(ctx));
-                self.patch(ctx, skip, op);
-                self.add_op(ctx, expression.span, Op::Pop);
-
-                self.expression(b, ctx);
-                let op = Op::Jmp(self.next_ip(ctx));
-                self.patch(ctx, jump, op);
-            }
-            Not(a) => self.un_op(a, &[Op::Not], expression.span, ctx),
-
-            Duplicate(a) => self.un_op(a, &[Op::Copy(1)], expression.span, ctx),
-
-            Parenthesis(expr) => self.expression(expr, ctx),
-
-            IfExpression {
-                condition,
-                pass,
-                fail,
-            } => {
-                self.expression(condition, ctx);
-
-                let skip = self.add_op(ctx, expression.span, Op::Illegal);
-
-                self.expression(pass, ctx);
-                let out = self.add_op(ctx, expression.span, Op::Illegal);
-                let op = Op::JmpFalse(self.next_ip(ctx));
-                self.patch(ctx, skip, op);
-
-                self.expression(fail, ctx);
-                let op = Op::Jmp(self.next_ip(ctx));
-                self.patch(ctx, out, op);
-            }
-
-            IfShort {
-                condition,
-                fail,
-                ..
-            } => {
-                // Results in 2 values pushed on the stack - since there's a Duplicate in
-                // there!
-                self.expression(condition, ctx);
-
-                let skip = self.add_op(ctx, expression.span, Op::Illegal);
-                let out = self.add_op(ctx, expression.span, Op::Illegal);
-
-                let op = Op::JmpFalse(self.next_ip(ctx));
-                self.patch(ctx, skip, op);
-
-                self.add_op(ctx, expression.span, Op::Pop);
-                self.expression(fail, ctx);
-                let op = Op::Jmp(self.next_ip(ctx));
-                self.patch(ctx, out, op);
-            }
-
-
-            Function {
-                name,
-                params,
-                ret,
-                body,
-            } => {
-                let file = self.file_from_namespace(ctx.namespace).display();
-                let name = format!("fn {} {}:{}", name, file, expression.span.line);
-
-                // === Frame begin ===
-                let inner_ctx = self.push_frame_and_block(ctx, &name, expression.span);
-                let mut param_types = Vec::new();
-                for (ident, ty) in params.iter() {
-                    param_types.push(self.resolve_type(&ty, inner_ctx));
-                    let param = self.define(&ident.name, VarKind::Const, ident.span);
-                    self.activate(param);
-                }
-                let ret = self.resolve_type(&ret, inner_ctx);
-                let ty = Type::Function(param_types, Box::new(ret));
-                self.blocks[inner_ctx.block_slot].ty = ty.clone();
-
-                self.statement(&body, inner_ctx);
-
-                if !all_paths_return(&body) {
-                    let nil = self.constant(Value::Nil);
-                    self.add_op(inner_ctx, body.span, nil);
-                    self.add_op(inner_ctx, body.span, Op::Return);
-                }
-
-                self.blocks[inner_ctx.block_slot].upvalues = self
-                    .pop_frame(inner_ctx)
-                    .upvalues
-                    .into_iter()
-                    .map(|u| (u.parent, u.upupvalue, u.ty))
-                    .collect();
-                let function = Value::Function(Rc::new(Vec::new()), ty, inner_ctx.block_slot);
-                // === Frame end ===
-
-                let function = self.constant(function);
-                self.add_op(ctx, expression.span, function);
-            }
-
-            Blob { blob, fields } => {
-                self.assignable(blob, ctx);
-                for (name, field) in fields.iter() {
-                    let name = self.constant(Value::String(Rc::new(name.clone())));
-                    self.add_op(ctx, field.span, name);
-                    self.expression(field, ctx);
-                }
-                self.add_op(ctx, expression.span, Op::Call(fields.len() * 2));
-            }
-
-            Tuple(x) | List(x) | Set(x) | Dict(x) => {
-                for expr in x.iter() {
-                    self.expression(expr, ctx);
-                }
-                self.add_op(
-                    ctx,
-                    expression.span,
-                    match &expression.kind {
-                        Tuple(_) => Op::Tuple(x.len()),
-                        List(_) => Op::List(x.len()),
-                        Set(_) => Op::Set(x.len()),
-                        Dict(_) => Op::Dict(x.len()),
-                        _ => unreachable!(),
-                    },
-                );
-            }
-
-            Float(a) => self.push(Value::Float(*a), expression.span, ctx),
-            Bool(a) => self.push(Value::Bool(*a), expression.span, ctx),
-            Int(a) => self.push(Value::Int(*a), expression.span, ctx),
-            Str(a) => self.push(Value::String(Rc::new(a.clone())), expression.span, ctx),
-            Nil => self.push(Value::Nil, expression.span, ctx),
-        }
-    }
-
     fn resolve_and_capture(&mut self, name: &str, frame: usize, span: Span) -> Result<Lookup, ()> {
         // Frame 0 has globals which cannot be captured.
         if frame == 0 {
@@ -585,94 +276,6 @@ impl Compiler {
         };
         let up = self.upvalue(up, frame);
         Ok(Lookup::Upvalue(up))
-    }
-
-    fn read_identifier(
-        &mut self,
-        name: &str,
-        span: Span,
-        ctx: Context,
-        namespace: usize,
-    ) -> Option<usize> {
-        match self.resolve_and_capture(name, ctx.frame, span) {
-            Ok(Lookup::Upvalue(up)) => {
-                let op = Op::ReadUpvalue(up.slot);
-                self.add_op(ctx, span, op);
-            }
-
-            Ok(Lookup::Variable(var)) => {
-                let op = Op::ReadLocal(var.slot);
-                self.add_op(ctx, span, op);
-            }
-
-            Err(()) => match self.namespaces[namespace].get(name) {
-                Some(Name::Global(slot)) => {
-                    let op = Op::ReadGlobal(*slot);
-                    self.add_op(ctx, span, op);
-                }
-                Some(Name::Blob(blob)) => {
-                    let op = Op::Constant(*blob);
-                    self.add_op(ctx, span, op);
-                }
-                Some(Name::Namespace(new_namespace)) => return Some(*new_namespace),
-                None => {
-                    if let Some((slot, _, _)) = self.functions.get(name) {
-                        let slot = *slot;
-                        let op = self.constant(Value::ExternFunction(slot));
-                        self.add_op(ctx, span, op);
-                    } else {
-                        error!(
-                            self,
-                            ctx,
-                            span,
-                            "Cannot read '{}' in '{}'",
-                            name,
-                            self.file_from_namespace(namespace).display()
-                        );
-                    }
-                }
-            },
-        }
-        None
-    }
-
-    fn set_identifier(&mut self, name: &str, span: Span, ctx: Context, namespace: usize) {
-        match self.resolve_and_capture(name, ctx.frame, span) {
-            Ok(Lookup::Upvalue(up)) => {
-                let op = Op::AssignUpvalue(up.slot);
-                self.add_op(ctx, span, op);
-                return;
-            }
-
-            Ok(Lookup::Variable(var)) => {
-                let op = Op::AssignLocal(var.slot);
-                self.add_op(ctx, span, op);
-                return;
-            }
-
-            Err(()) => match self.namespaces[namespace].get(name) {
-                Some(Name::Global(slot)) => {
-                    let var = &self.frames[0].variables[*slot];
-                    if var.kind.immutable() && ctx.frame != 0 {
-                        error!(self, ctx, span, "Cannot mutate constant '{}'", name);
-                    } else {
-                        let op = Op::AssignGlobal(var.slot);
-                        self.add_op(ctx, span, op);
-                    }
-                }
-
-                _ => {
-                    error!(
-                        self,
-                        ctx,
-                        span,
-                        "Cannot assign '{}' in '{}'",
-                        name,
-                        self.file_from_namespace(namespace).display()
-                    );
-                }
-            },
-        }
     }
 
     fn resolve_type_namespace(
@@ -841,234 +444,6 @@ impl Compiler {
         self.frames.last_mut().unwrap().variables[slot].active = true;
     }
 
-    fn statement(&mut self, statement: &Statement, ctx: Context) {
-        use StatementKind::*;
-        self.panic = false;
-
-        match &statement.kind {
-            Use { .. } | EmptyStatement => {}
-
-            Blob { name, fields } => {
-                let fields = fields.iter()
-                    .map(|(k, v)| (k.clone(), self.resolve_type(&v, ctx)))
-                    .collect();
-                if let Some(Name::Blob(slot)) = self.namespaces[ctx.namespace].get(name) {
-                    match &mut self.constants[*slot] {
-                        Value::Ty(Type::Blob(_, b_fields)) => {
-                            *b_fields = fields;
-                        }
-                        _ => unreachable!(),
-                    }
-                } else {
-                    error!(
-                        self,
-                        ctx,
-                        statement.span,
-                        "No blob with the name '{}' in this namespace (#{})",
-                        name,
-                        ctx.namespace
-                    );
-                }
-            }
-
-            IsCheck { lhs, rhs } => {
-                let lhs = self.resolve_type(lhs, ctx);
-                let rhs = self.resolve_type(rhs, ctx);
-                if let Err(msg) = rhs.fits(&lhs) {
-                    error!(
-                        self,
-                        ctx, statement.span,
-                        "Is-check failed - {}",
-                        msg
-                    );
-                }
-            }
-
-            #[rustfmt::skip]
-            Definition { ident, kind, value, .. } => {
-                // TODO(ed): Don't use type here - type check the tree first.
-                self.expression(value, ctx);
-
-                if ctx.frame == 0 {
-                    // Global
-                    self.set_identifier(&ident.name, statement.span, ctx, ctx.namespace);
-                } else {
-                    // Local variable
-                    let slot = self.define(&ident.name, *kind, statement.span);
-                    self.activate(slot);
-                }
-            }
-
-            #[rustfmt::skip]
-            Assignment { target, value, kind } => {
-                use AssignableKind::*;
-                use ParserOp::*;
-
-                let mutator = |kind| matches!(kind, Add | Sub | Mul | Div);
-
-                let write_mutator_op = |comp: &mut Self, ctx, kind| {
-                    let op = match kind {
-                        Add => Op::Add,
-                        Sub => Op::Sub,
-                        Mul => Op::Mul,
-                        Div => Op::Div,
-                        Nop => {
-                            return;
-                        }
-                    };
-                    comp.add_op(ctx, statement.span, op);
-                };
-
-                match &target.kind {
-                    Read(ident) => {
-                        if mutator(*kind) {
-                            self.read_identifier(&ident.name, statement.span, ctx, ctx.namespace);
-                        }
-                        self.expression(value, ctx);
-
-                        write_mutator_op(self, ctx, *kind);
-
-                        self.set_identifier(&ident.name, statement.span, ctx, ctx.namespace);
-                    }
-                    ArrowCall(..) | Call(..) => {
-                        error!(
-                            self,
-                            ctx, statement.span, "Cannot assign to result from function call"
-                        );
-                    }
-                    Access(a, field) => {
-                        if let Some(namespace) = self.assignable(a, ctx) {
-                            if mutator(*kind) {
-                                self.read_identifier(&field.name, statement.span, ctx, namespace);
-                            }
-                            self.expression(value, ctx);
-
-                            write_mutator_op(self, ctx, *kind);
-
-                            self.set_identifier(&field.name, statement.span, ctx, namespace);
-                        } else {
-                            let slot = self.string(&field.name);
-
-                            if mutator(*kind) {
-                                self.add_op(ctx, statement.span, Op::Copy(1));
-                                self.add_op(ctx, field.span, Op::GetField(slot));
-                            }
-                            self.expression(value, ctx);
-
-                            write_mutator_op(self, ctx, *kind);
-
-                            self.add_op(ctx, field.span, Op::AssignField(slot));
-                        }
-                    }
-                    Index(a, b) => {
-                        self.assignable(a, ctx);
-                        self.expression(b, ctx);
-
-                        if mutator(*kind) {
-                            self.add_op(ctx, statement.span, Op::Copy(2));
-                            self.add_op(ctx, statement.span, Op::GetIndex);
-                        }
-                        self.expression(value, ctx);
-                        write_mutator_op(self, ctx, *kind);
-
-                        self.add_op(ctx, statement.span, Op::AssignIndex);
-                    }
-                    Expression(_) => {
-                        unreachable!("Cannot assign to 'AssignableKind::Expression'");
-                    }
-                }
-            }
-
-            StatementExpression { value } => {
-                self.expression(value, ctx);
-                self.add_op(ctx, statement.span, Op::Pop);
-            }
-
-            Block { statements } => {
-                let stack_size = self.frames[ctx.frame].variables.len();
-
-                for statement in statements {
-                    self.statement(statement, ctx);
-                }
-
-                self.pop_until_size(ctx, statement.span, stack_size);
-            }
-
-            Loop { condition, body } => {
-                let start = self.next_ip(ctx);
-                self.expression(condition, ctx);
-                let jump_from = self.add_op(ctx, condition.span, Op::Illegal);
-
-                // Skip the next op - it's for the break
-                //  start: .. condition ..
-                //         JmpFalse(over)
-                //  break: Jmp(end)
-                //   over: .. loop ..
-                //    end: ..
-                self.add_op(ctx, condition.span, Op::Jmp(jump_from + 3));
-                let break_from = self.add_op(ctx, condition.span, Op::Illegal);
-
-                let stack_size = self.frames[ctx.frame].variables.len();
-                self.loops.push(LoopFrame {
-                    continue_addr: start,
-                    break_addr: break_from,
-                    stack_size,
-                });
-                self.statement(body, ctx);
-                self.loops.pop();
-
-                self.add_op(ctx, body.span, Op::Jmp(start));
-                let out = self.next_ip(ctx);
-                self.patch(ctx, jump_from, Op::JmpFalse(out));
-                self.patch(ctx, break_from, Op::Jmp(out));
-            }
-
-            #[rustfmt::skip]
-            If { condition, pass, fail } => {
-                self.expression(condition, ctx);
-
-                let jump_from = self.add_op(ctx, condition.span, Op::Illegal);
-                self.statement(pass, ctx);
-                let jump_out = self.add_op(ctx, condition.span, Op::Illegal);
-                self.statement(fail, ctx);
-
-                self.patch(ctx, jump_from, Op::JmpFalse(jump_out + 1));
-
-                let end = self.next_ip(ctx);
-                self.patch(ctx, jump_out, Op::Jmp(end));
-            }
-
-            Continue {} => match self.loops.last().cloned() {
-                Some(LoopFrame { stack_size, continue_addr, .. }) => {
-                    self.emit_pop_until_size(ctx, statement.span, stack_size);
-                    self.add_op(ctx, statement.span, Op::Jmp(continue_addr));
-                }
-                None => {
-                    error!(self, ctx, statement.span, "`continue` statement not in a loop");
-                }
-            }
-
-            Break {} => match self.loops.last().cloned() {
-                Some(LoopFrame { stack_size, break_addr, .. }) => {
-                    self.emit_pop_until_size(ctx, statement.span, stack_size);
-                    self.add_op(ctx, statement.span, Op::Jmp(break_addr));
-                }
-                None => {
-                    error!(self, ctx, statement.span, "`continue` statement not in a loop");
-                }
-            }
-
-            Unreachable {} => {
-                self.add_op(ctx, statement.span, Op::Unreachable);
-            }
-
-            Ret { value } => {
-                self.expression(value, ctx);
-                self.add_op(ctx, statement.span, Op::Return);
-            }
-        }
-    }
-
     fn compile(
         mut self,
         typecheck: bool,
@@ -1077,13 +452,10 @@ impl Compiler {
     ) -> Result<Prog, Vec<Error>> {
         assert!(!tree.modules.is_empty(), "Cannot compile an empty program");
         let name = "/preamble/";
-        let mut block = Block::new(name, 0, &tree.modules[0].0);
-        block.ty = Type::Function(Vec::new(), Box::new(Type::Void));
-        self.blocks.push(block);
-        self.frames.push(Frame::new(name, Span::zero()));
-        let mut ctx = Context {
-            block_slot: self.blocks.len() - 1,
-            frame: self.frames.len() - 1,
+        let start_span = tree.modules[0].1.span;
+        self.frames.push(Frame::new(name, start_span));
+        let ctx = Context {
+            frame: 0,
             ..Context::from_namespace(0)
         };
 
@@ -1117,10 +489,15 @@ impl Compiler {
             return Err(self.errors);
         }
 
+        let mut bytecode_compiler = bytecode::BytecodeCompiler::new(&mut self);
+        bytecode_compiler.preamble(start_span);
+
         for (statement, namespace) in statements.iter() {
-            ctx.namespace = *namespace;
-            self.statement(statement, ctx);
+            bytecode_compiler.compile(statement, *namespace);
         }
+
+        bytecode_compiler.postamble(start_span);
+        let blocks = bytecode_compiler.blocks;
 
         if !self.errors.is_empty() {
             return Err(self.errors);
@@ -1130,20 +507,10 @@ impl Compiler {
             typechecker::solve(&mut self, &statements)?;
         }
 
-        self.read_identifier("start", Span::zero(), ctx, 0);
-        self.add_op(ctx, Span::zero(), Op::Call(0));
-
-        let module = &tree.modules[0].1;
-        let nil = self.constant(Value::Nil);
-        self.add_op(ctx, module.span, nil);
-        self.add_op(ctx, module.span, Op::Return);
-
-        self.pop_frame(ctx);
 
         if self.errors.is_empty() {
             Ok(Prog {
-                blocks: self
-                    .blocks
+                blocks: blocks
                     .into_iter()
                     .map(|x| Rc::new(RefCell::new(x)))
                     .collect(),
@@ -1200,13 +567,9 @@ impl Compiler {
                         let other = path_to_namespace_id[file];
                         (Name::Namespace(other), ident.name.clone(), ident.span)
                     }
-                    Definition { ident: Identifier { name, span }, kind, .. } => {
+                    Definition { ident: Identifier { name, .. }, kind, .. } => {
                         let var = self.define(name, *kind, statement.span);
                         self.activate(var);
-
-                        // Uninitalized values have the value Nil
-                        let unknown = self.constant(Value::Nil);
-                        self.add_op(ctx, *span, unknown);
 
                         (Name::Global(var), name.clone(), statement.span)
                     }
@@ -1259,28 +622,6 @@ fn parse_signature(func_name: &str, sig: &str) -> ParserType {
 
 pub fn compile(typecheck: bool, prog: AST, functions: &[(String, RustFunction, String)]) -> Result<Prog, Vec<Error>> {
     Compiler::new().compile(typecheck, prog, functions)
-}
-
-fn all_paths_return(statement: &Statement) -> bool {
-    match &statement.kind {
-        StatementKind::Use { .. }
-        | StatementKind::Blob { .. }
-        | StatementKind::IsCheck { .. }
-        | StatementKind::Assignment { .. }
-        | StatementKind::Definition { .. }
-        | StatementKind::StatementExpression { .. }
-        | StatementKind::Unreachable
-        | StatementKind::Continue
-        | StatementKind::Break
-        | StatementKind::EmptyStatement => false,
-
-        StatementKind::If { pass, fail, .. } => all_paths_return(pass) && all_paths_return(fail),
-
-        StatementKind::Loop { body, .. } => all_paths_return(body),
-        StatementKind::Block { statements } => statements.iter().any(all_paths_return),
-
-        StatementKind::Ret { .. } => true,
-    }
 }
 
 pub(crate) fn first_ok_or_errs<I, T, E>(mut iter: I) -> Result<T, Vec<E>>
